requirements.txt
Numpy
PyOpenGl
PyOpenGL_accelerate
glfw